src/bot/requirements.txt
pyTelegramBotAPI==4.14.0
requests==2.32.3
python-dotenv==1.0.0 